requirements.txt
aiohttp==3.6.2
aliyun-python-sdk-core==2.13.35
aliyun-python-sdk-core-v3==2.13.32
aliyun-python-sdk-kms==2.15.0
amqp==2.6.1
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.5.3
billiard==3.6.4.0
celery==4.4.0
certifi==2021.5.30
cffi==1.14.6
chardet==3.0.4
click==7.1.2
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
crcmod==1.7
cryptography==3.4.7
Django==3.0.8
django-celery-beat==1.5.0
django-celery-results==1.2.1
django-cors-headers==3.4.0
django-crontab==0.7.1
django-debug-toolbar==2.2
django-filter==2.3.0
django-redis==4.12.1
django-rest-swagger==2.2.0
django-softdelete-new==0.2.1
django-timezone-field==4.2.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
dnspython==2.1.0
drf-extensions==0.6.0
drf-renderer-xlsx==0.3.7
drf-yasg==1.17.1
emoji==0.5.4
et-xmlfile==1.1.0
eventlet==0.25.2
flex==6.14.1
flower==0.9.4
gevent==20.5.0
greenlet==1.1.1
h11==0.9.0
httpie==2.2.0
humanize==0.5.1
idna==2.10
idna-ssl==1.1.0
importlib-metadata==4.6.4
inflection==0.5.1
itypes==1.2.0
jieba==0.42
Jinja2==3.0.1
jmespath==0.10.0
jsonpointer==2.1
jsonschema==3.2.0
kombu==4.6.11
MarkupSafe==2.0.1
monotonic==1.6
msgpack==1.0.0
multidict==4.7.6
mysqlclient==2.0.1
numpy==1.19.5
openapi-codec==1.3.2
openpyxl==3.0.7
oss2==2.12.1
packaging==21.0
pandas==1.1.5
pika==1.1.0
Pillow==7.2.0
pycodestyle==2.7.0
pycparser==2.20
pycryptodome==3.10.1
Pygments==2.10.0
PyJWT==1.7.1
PyMySQL==0.9.3
pyotp==2.3.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-crontab==2.5.1
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
requests==2.24.0
rfc3987==1.3.8
ruamel.yaml==0.17.13
ruamel.yaml.clib==0.2.6
simplejson==3.17.4
six==1.16.0
sqlparse==0.4.1
strict-rfc3339==0.7
swagger-spec-validator==2.7.3
toml==0.10.2
tornado==6.0.4
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.25.11
uvicorn==0.11.5
validate-email==1.3
vine==1.3.0
websockets==8.1
xlrd==1.2.0
xlwt==1.3.0
yarl==1.6.3
zipp==3.5.0
